update gsm util

+++ utils/gsm/gsm.sh
@@ -1,15 +1,23 @@
 #!/bin/bash
 PWRKEY=6
 RESET=7
-
-echo $PWRKEY > /sys/class/gpio/export
-echo $RESET > /sys/class/gpio/export
-
 PWRKEY_GPIO=/sys/class/gpio/gpio${PWRKEY}
 RESET_GPIO=/sys/class/gpio/gpio${RESET}
 
 
+function init() {
+	if [ ! -e  $PWRKEY_GPIO ]; then
+		echo $PWRKEY > /sys/class/gpio/export
+	fi
+
+	if [ ! -e  $RESET_GPIO ]; then
+		echo $RESET > /sys/class/gpio/export
+	fi
+}
+
+
 function toggle() {
+	echo "toggle SIM900 state using PWRKEY"
 	echo 0 > ${PWRKEY_GPIO}/value
 	sleep 1
 	echo 1 > ${PWRKEY_GPIO}/value
@@ -17,8 +25,11 @@ function toggle() {
 	echo 0 > ${PWRKEY_GPIO}/value
 }
 
+
+init
+
 echo out > ${PWRKEY_GPIO}/direction
-echo out   > ${RESET_GPIO}direction
+echo out   > ${RESET_GPIO}/direction
 
 echo 0 > ${RESET_GPIO}/value
 
@@ -32,11 +43,10 @@ case "$1" in
 	;;
 
 	"toggle" )
-		echo "toggle SIM900 state using PWRKEY"
 		toggle
 	;;
 	* )
-		echo "USAGE: $0 [toggle|reset]\n toggle by default";
+		echo "USAGE: $0 toggle|<reset>";
 		toggle
 	;;
 esac
